requirements.txt
numpy>=1.24.4
torch>=2.0.0
torchvision
matplotlib
pandas
datasets
transformers
seaborn
scikit-learn
python-Levenshtein
nltk
rouge
fastapi
uvicorn
torchmetrics